+++ src/main/java/wooteco/subway/dto/request/StationRequest.java
@@ -1,6 +1,12 @@
 package wooteco.subway.dto.request;
 
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
+
 public class StationRequest {
+
+    @NotBlank
+    @Pattern(regexp = "^[가-힣|A-Z|a-z| 0-9]*역$")
     private String name;
 
     public StationRequest() {
