+++ src/main/java/wooteco/subway/repository/dao/StationDao.java
@@ -1,10 +1,10 @@
-package wooteco.subway.dao;
+package wooteco.subway.repository.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.station.Station;
+import wooteco.subway.domain.Station;
 
 import java.util.HashMap;
 import java.util.List;
