+++ src/main/java/wooteco/subway/service/StationService.java
@@ -1,10 +1,10 @@
 package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.station.Station;
+import wooteco.subway.domain.Station;
 import wooteco.subway.dto.response.StationResponse;
 import wooteco.subway.exception.station.StationDuplicateException;
+import wooteco.subway.repository.dao.StationDao;
 
 import java.util.List;
 import java.util.stream.Collectors;
