+++ src/test/java/wooteco/subway/line/LineDaoTest.java
@@ -7,9 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.annotation.Rollback;
-import wooteco.subway.dao.LineDao;
-import wooteco.subway.domain.line.Line;
+import wooteco.subway.domain.Line;
+import wooteco.subway.repository.dao.LineDao;
 
 import java.util.List;
 
@@ -17,7 +16,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @JdbcTest
-@Rollback
 class LineDaoTest {
     LineDao lineDao;
