+++ src/test/java/wooteco/subway/station/StationDaoTest.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.station.Station;
+import wooteco.subway.domain.Station;
+import wooteco.subway.repository.dao.StationDao;
 
 import java.util.List;
